+++ src/compiler/compile/render_ssr/handlers/Element.ts
@@ -157,10 +157,6 @@ export default function (node: Element, renderer: Renderer, options: RenderOptio
 		}
 	});
 
-	if (options.hydratable && options.head_id) {
-		renderer.add_string(` data-svelte="${options.head_id}"`);
-	}
-
 	renderer.add_string('>');
 
 	if (node_contents !== undefined) {