+++ test/runtime/samples/dynamic-element-void-with-content-5/_config.js
@@ -1,0 +1,6 @@
+export default {
+	compileOptions: {
+		dev: true
+	},
+	warnings: ['<svelte:element this="input"> is self-closing and cannot have content.']
+};