+++ test/validator/samples/a11y-aria-proptypes-token/input.svelte
@@ -1,0 +1,6 @@
+<div aria-sort="" />
+<div aria-sort="descnding" />
+<div aria-sort />
+<div aria-sort={true} />
+<div aria-sort={"false"} />
+<div aria-sort="ascending descending" />