+++ test/validator/samples/a11y-aria-proptypes-tokenlist/input.svelte
@@ -1,0 +1,7 @@
+<div aria-relevant="" />
+<div aria-relevant="foobar" />
+<div aria-relevant />
+<div aria-relevant={true} />
+<div aria-relevant={"false"} />
+<div aria-relevant="additions removalss" />
+<div aria-relevant="additions removalss " />